backend/hub/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-decouple>=3.8
gunicorn>=21.2
whitenoise>=6.6
psycopg2-binary>=2.9
pyodbc>=5.0
sentence-transformers>=2.7
scikit-learn>=1.4
numpy>=1.26
